Updated shell script to make and update pot and po files

+++ i18n/make_template.sh
@@ -11,6 +11,19 @@ then
     JOIN="-j"
 fi
 
-find $PROJECT_DIR -iname "*.php" | xargs xgettext -L PHP --keyword=l:1 -F --copyright-holder=PoliEdro $JOIN -d messages -p $DIR --from-code=UTF-8
+find $PROJECT_DIR -iname "*.php" > $DIR/phps
 
-mv $DIR/messages.po $DIR/messages.pot
+if [ -f $DIR/phps ]
+then
+  xgettext -L PHP --keyword=l:1 -F --copyright-holder=PoliEdro $JOIN -d messages -p $DIR --from-code=UTF-8 -f $DIR/phps
+
+  mv $DIR/messages.po $DIR/messages.pot
+  rm $DIR/phps
+
+  for i in $(find $DIR -iname "*.po");
+  do
+    msgmerge -U $i $DIR/messages.pot
+    echo "Updated file $i "
+  done
+
+fi
